+++ crates/sword-macros/src/controllers/mod.rs
@@ -1,7 +1,7 @@
 pub mod shared;
 
 #[cfg(feature = "event-handlers")]
-pub mod mem_event_handler;
+pub mod event_handler;
 
 #[cfg(feature = "grpc-controllers")]
 pub mod grpc;
@@ -39,8 +39,8 @@ pub fn expand_controller(attr: TokenStream, item: TokenStream) -> syn::Result<To
         ParsedControllerKind::Grpc { .. } => grpc::expand_grpc_controller(&parsed_input)?,
 
         #[cfg(feature = "event-handlers")]
-        ParsedControllerKind::MemEventHandler { .. } => {
-            mem_event_handler::expand_mem_event_handler(&parsed_input)?
+        ParsedControllerKind::EventHandler { .. } => {
+            event_handler::expand_event_handler(&parsed_input)?
         }
     };